% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheapr.R
\docType{package}
\name{cheapr-package}
\alias{cheapr}
\alias{cheapr-package}
\title{cheapr: Simple Functions to Save Time and Memory}
\description{
In this package, 'cheap' means fast and efficient.

cheapr aims to provide a set of functions for programmers to write
cheaper code, saving time and memory.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/NicChr/cheapr/issues}
}

}
\author{
\strong{Maintainer}: Nick Christofides \email{nick.christofides.r@gmail.com} (\href{https://orcid.org/0000-0002-9743-7342}{ORCID})

}
