% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine.R
\name{cheapr_c}
\alias{cheapr_c}
\title{A cheapr version of \code{c()}}
\usage{
cheapr_c(...)
}
\arguments{
\item{...}{Objects to combine.}
}
\value{
Combined objects.
}
\description{
cheapr's version of \code{c()}. It is quite a bit faster for atomic vectors
and combines data frame rows instead of cols.
}
