% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheapr_table.R
\name{cheapr_table}
\alias{cheapr_table}
\title{Fast frequency tables - Still experimental}
\usage{
cheapr_table(
  ...,
  names = TRUE,
  order = FALSE,
  na_exclude = FALSE,
  classed = FALSE
)
}
\arguments{
\item{...}{\verb{>=1} objects that can be converted to a factor through
\code{cheapr::factor_()}.}

\item{names}{Should level names be kept? Default is \code{TRUE}.}

\item{order}{Should result be ordered by level names? Default is
\code{FALSE}.}

\item{na_exclude}{Should \code{NA} values be excluded? Default is \code{FALSE}.}

\item{classed}{Should a \code{table} object be returned? Default is \code{FALSE}}
}
\value{
A named integer vector if one object is supplied, otherwise an
array.
}
\description{
This is not a 1-to-1 copy of \code{base::table()} as some behaviours differ.
It is more flexible as it accepts inputs such as data frames and
\code{vctrs_rcrd} objects.
}
