% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{cheapr_table}
\alias{cheapr_table}
\alias{counts}
\alias{table_}
\title{Fast frequency tables - Still experimental}
\usage{
cheapr_table(
  ...,
  names = TRUE,
  order = FALSE,
  na_exclude = FALSE,
  classed = FALSE
)

counts(x, sort = is.factor(x))

table_(..., names = TRUE, order = FALSE, na_exclude = FALSE, classed = FALSE)
}
\arguments{
\item{...}{\verb{>=1} objects that can be converted to a factor through
\code{cheapr::factor_()}.}

\item{names}{Should level names be kept? Default is \code{TRUE}.}

\item{order}{Should result be ordered by level names? Default is
\code{FALSE}.}

\item{na_exclude}{Should \code{NA} values be excluded? Default is \code{FALSE}.}

\item{classed}{Should a \code{table} object be returned? Default is \code{FALSE}}

\item{x}{A vector.}

\item{sort}{Should groups be sorted? Default is \code{FALSE}.}
}
\value{
A named integer vector if one object is supplied, otherwise an
array.
}
\description{
This is not a one-to-one copy of \code{base::table()} as some behaviours differ.
It is more flexible as it accepts inputs such as data frames and
\code{vctrs_rcrd} objects.
}
\details{
\code{cheapr_table()} tries to match the behaviour of \code{table()} where possible.
\code{counts()} is an alternative that returns a \code{data.frame} of unique keys and
counts.
}
