% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_sset.R
\name{df_sset}
\alias{df_sset}
\alias{df_slice}
\alias{df_select}
\title{Fast functions for data frame subsetting}
\usage{
df_sset(x, i = NULL, j = NULL, keep_attrs = FALSE)

df_slice(x, i = NULL)

df_select(x, j = NULL)
}
\arguments{
\item{x}{A \code{data.frame}.}

\item{i}{Rows - If \code{NULL} all rows are returned.}

\item{j}{Cols - If \code{NULL} all cols are returned.}

\item{keep_attrs}{Should all attributes (except for \code{names} and \code{row.names})
be kept as is? The default is \code{FALSE} which returns a plain data frame.}
}
\value{
A data frame subsetted on rows \code{i} and cols \code{j}.
}
\description{
These functions are for developers that need minimal overhead when
filtering on rows and/or cols.
}
