% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{named_list}
\alias{named_list}
\title{Turn dot-dot-dot (\code{...}) into a named list}
\usage{
named_list(..., .keep_null = TRUE)
}
\arguments{
\item{...}{Key-value pairs.}

\item{.keep_null}{Should \code{NULL} entries be kept? Default is \code{TRUE}.}
}
\value{
A named list.
}
\description{
A fast and useful function for always returning a named list from \code{...}
}
