% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_sign.R
\name{int_sign}
\alias{int_sign}
\title{A fast and integer-based \code{sign()}}
\usage{
int_sign(x)
}
\arguments{
\item{x}{Integer or double vector.}
}
\value{
An integer vector denoting the sign, -1 for negatives, 1 for positives and 0
for when \code{x == 0}.
}
\description{
A fast and integer-based \code{sign()}
}
