% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{math}
\alias{math}
\alias{abs_}
\alias{floor_}
\alias{ceiling_}
\alias{trunc_}
\alias{negate_}
\alias{exp_}
\alias{sqrt_}
\alias{sign_}
\alias{log_}
\alias{log10_}
\alias{round_}
\title{Parallelised math operations}
\usage{
abs_(x)

floor_(x)

ceiling_(x)

trunc_(x)

negate_(x)

exp_(x)

sqrt_(x)

sign_(x)

log_(x, base = exp(1))

log10_(x)

round_(x, digits = 0)
}
\arguments{
\item{x}{\verb{[numeric(n)]} vector.}

\item{base}{\verb{[numeric(n)]} - Logarithm base.}

\item{digits}{\verb{[numeric(n)]} - Number of digits to round to.}

\item{y}{\verb{[numeric(n)]} vector.}
}
\value{
A transformed integer or double vector.
}
\description{
Parallelised math operations
}
