% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_init.R
\name{na_init}
\alias{na_init}
\title{Fast \code{NA} initialisation}
\usage{
na_init(x, n = 0L)
}
\arguments{
\item{x}{A vector.}

\item{n}{Vector length to initialise.}
}
\value{
Initialises \code{NA} values of the same type as \code{x}.
}
\description{
Fast \code{NA} initialisation
}
\seealso{
\link{rep_len_}
}
