% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{paste_}
\alias{paste_}
\alias{paste0_}
\title{Fast string concatenation using C++}
\usage{
paste_(..., .sep = " ", .args = NULL)

paste0_(..., .args = NULL)
}
\arguments{
\item{...}{Character vectors to concatenate.}

\item{.sep}{\verb{[character(1)]} - A string to separate the supplied strings.}

\item{.args}{An alternative to \code{...} so you can supply arguments directly
in a list. \cr
This is equivalent to \code{do.call(f, .args)} but much more efficient.}
}
\description{
Fast string concatenation using C++
}
\examples{
library(cheapr)

# Normal usage
paste_("Hello", "and", "Goodbye", sep = " ")
paste0_(100, "\%")

# Recycling with zero-length vectors
paste_("hello", character(), letters)

# Concatenating many objects is very fast via `.args`
strings <- sample_(letters, 5e04, TRUE) |>
  with_local_seed(1) |>
  as.list()
mark(paste_(.args = strings), do.call(paste, strings))

}
