% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_math.R
\name{set_abs}
\alias{set_abs}
\alias{set_floor}
\alias{set_ceiling}
\alias{set_trunc}
\alias{set_exp}
\alias{set_sqrt}
\alias{set_change_sign}
\alias{set_round}
\alias{set_log}
\title{Math operations by reference}
\usage{
set_abs(x)

set_floor(x)

set_ceiling(x)

set_trunc(x)

set_exp(x)

set_sqrt(x)

set_change_sign(x)

set_round(x, digits = 0)

set_log(x, base = exp(1))
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Number of digits to round to.}

\item{base}{Logarithm base.}
}
\value{
The exact same object with no copy made, just transformed.
}
\description{
These functions transform your variable by reference, with no copies being made.
}
\details{
These functions are particularly useful for situations
where you have made a copy and then
wish to perform further operations without creating more copies. \cr
\code{NA} and \code{NaN} values are ignored and never updated. \cr
These functions will \bold{not work} on \bold{any} classed objects, meaning they
only work on standard integer and numeric vectors and matrices.
}
\examples{
library(cheapr)
library(bench)

x <- rnorm(2e05)
options(cheapr.cores = 2)
mark(
  base = exp(log(abs(x))),
  cheapr = set_exp(set_log(set_abs(x)))
)
options(cheapr.cores = 1)

}
