% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threads.R
\name{get_max_threads}
\alias{get_max_threads}
\alias{set_threads}
\alias{get_threads}
\title{Get and set the number of OpenMP threads to be used in cheapr}
\usage{
get_max_threads()

set_threads(n)

get_threads()
}
\arguments{
\item{n}{\verb{[integer(1)]} - Number of threads to use.}
}
\value{
\code{get_max_threads()} returns the max number of threads you can use. \cr
\code{get_threads()} returns the number of threads that are being used. \cr
\code{set_threads()} invisibly sets the number of threads to be used.
}
\description{
The default number of threads in
cheapr is 2 (if your system has at least 2 available threads).
You can change the number of threads via \code{set_threads()}.
To see the number of threads currently being used, run \code{get_threads()}.
To see the maximum number of threads your machine can use,
run \code{get_max_threads()}
}
