% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cast.R
\name{cast}
\alias{cast}
\alias{cast_common}
\title{Fast casting/coercing of R objects}
\usage{
cast(x, y)

cast_common(..., .args = NULL)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector.}

\item{...}{Vectors.}

\item{.args}{An alternative to \code{...} so you can supply arguments directly
in a list. \cr
This is equivalent to \code{do.call(f, .args)} but much more efficient.}
}
\value{
\code{cast()} coerces \code{x} into a common type between \code{x} and \code{y}.
\code{cast_common()} coerces all supplied vectors into a common type between them.
}
\description{
\code{cast()} is type-commutative, meaning the order \code{x} and \code{y} doesn't
affect the outcome type.
\code{cast()} will attempt to cast \code{x} into a common type between \code{x} and \code{y}.
}
