% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{named_dots}
\alias{named_dots}
\title{Turn dot-dot-dot (\code{...}) into a named list}
\usage{
named_dots(...)
}
\arguments{
\item{...}{Key-value pairs.}
}
\value{
A named list.
}
\description{
A fast and useful function for always returning a named list from \code{...}
}
