% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extras.R
\name{enframe_}
\alias{enframe_}
\alias{deframe_}
\alias{setdiff_}
\alias{intersect_}
\alias{cut_numeric}
\alias{\%in_\%}
\alias{\%!in_\%}
\title{Extra utilities}
\usage{
enframe_(x, name = "name", value = "value")

deframe_(x)

setdiff_(x, y, dups = TRUE)

intersect_(x, y, dups = TRUE)

cut_numeric(
  x,
  breaks,
  labels = NULL,
  include.lowest = FALSE,
  right = TRUE,
  dig.lab = 3L,
  ordered_result = FALSE,
  ...
)

x \%in_\% table

x \%!in_\% table
}
\arguments{
\item{x}{A vector or data frame.}

\item{name}{The column name to assign the names of a vector.}

\item{value}{The column name to assign the values of a vector.}

\item{y}{A vector or data frame.}

\item{dups}{Should duplicates be kept? Default is \code{TRUE}.}

\item{breaks}{See \code{?cut}.}

\item{labels}{See \code{?cut}.}

\item{include.lowest}{See \code{?cut}.}

\item{right}{See \code{?cut}.}

\item{dig.lab}{See \code{?cut}.}

\item{ordered_result}{See \code{?cut}.}

\item{...}{See \code{?cut}.}

\item{table}{See \code{?collapse::fmatch}}
}
\description{
Extra utilities
}
\details{
\code{intersect_()} and \code{setdiff_()} are faster and more efficient
alternatives to \code{intersect()} and \code{setdiff()} respectively. \cr
\code{enframe_()} and \code{deframe_()} are faster alternatives to
\code{tibble::enframe()} and \code{tibble::deframe()} respectively. \cr
\code{cut_numeric()} is a faster and more efficient alternative to
\code{cut.default()}.
}
