% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{abs_}
\alias{abs_}
\alias{floor_}
\alias{ceiling_}
\alias{trunc_}
\alias{negate_}
\alias{exp_}
\alias{sqrt_}
\alias{log_}
\alias{log10_}
\title{Parallelised math operations}
\usage{
abs_(x)

floor_(x)

ceiling_(x)

trunc_(x)

negate_(x)

exp_(x)

sqrt_(x)

log_(x, base = exp(1))

log10_(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Parallelised math operations
}
