% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_df.R
\name{new_df}
\alias{new_df}
\title{Fast data frame constructor}
\usage{
new_df(..., ..N = NULL, .recycle = FALSE, .name_repair = FALSE)
}
\arguments{
\item{...}{Key-value pairs.}

\item{..N}{\code{integer(1)} (Optional) number of rows. \cr
Commonly used to initialise a 0-column data frame with rows.}

\item{.recycle}{\code{logical(1)} Should arguments be recycled?
Default is \code{FALSE}.}

\item{.name_repair}{\code{logical(1)} Should duplicate names be made unique?
Default is \code{FALSE}.}
}
\value{
A \code{data.frame}
}
\description{
Fast data frame constructor
}
