% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview}
\alias{overview}
\alias{overview.default}
\alias{overview.logical}
\alias{overview.numeric}
\alias{overview.character}
\alias{overview.factor}
\alias{overview.Date}
\alias{overview.POSIXt}
\alias{overview.data.frame}
\title{An alternative to \code{summary()} inspired by the skimr package}
\usage{
overview(x, hist = FALSE)

\method{overview}{default}(x, hist = FALSE)

\method{overview}{logical}(x, hist = FALSE)

\method{overview}{numeric}(x, hist = FALSE)

\method{overview}{character}(x, hist = FALSE)

\method{overview}{factor}(x, hist = FALSE)

\method{overview}{Date}(x, hist = FALSE)

\method{overview}{POSIXt}(x, hist = FALSE)

\method{overview}{data.frame}(x, hist = FALSE)
}
\arguments{
\item{x}{A vector or data frame.}

\item{hist}{Should in-line histograms be returned? Default is \code{FALSE}.}
}
\value{
\code{overview(x)} returns a 1-row data frame unless
\code{x} is a data frame, in which case an object of class "overview" is returned,
Under the hood this is just a a list of data frames.
Key summary statistics are reported in each data frame.
}
\description{
A cheaper \code{summary()} function, designed for larger data.
}
