% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overview.R
\name{overview}
\alias{overview}
\title{An alternative to \code{summary()} inspired by the skimr package}
\usage{
overview(x, hist = FALSE)
}
\arguments{
\item{x}{A vector or data frame.}

\item{hist}{Should in-line histograms be returned? Default is \code{FALSE}.}
}
\value{
\code{overview(x)} returns a 1-row data frame unless
\code{x} is a data frame, in which a list of data frames is returned.
Key summary statistics are reported in each data frame.
}
\description{
A cheaper \code{summary()} function, designed for larger data.
}
