% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sset.R
\name{sset}
\alias{sset}
\alias{sset.data.frame}
\alias{sset.tbl_df}
\alias{sset.POSIXlt}
\alias{sset.data.table}
\alias{sset.sf}
\title{Cheaper subset}
\usage{
sset(x, i = 0, ...)

\method{sset}{data.frame}(x, i = 0, j = seq_along(x), ...)

\method{sset}{tbl_df}(x, i = 0, j = seq_along(x), ...)

\method{sset}{POSIXlt}(x, i = 0, ...)

\method{sset}{data.table}(x, i = 0, j = seq_along(x), ...)

\method{sset}{sf}(x, i = 0, j = seq_along(x), ...)
}
\arguments{
\item{x}{Vector or data frame.}

\item{i}{A logical or vector of indices. \cr
The default is 0 which differs to \code{[}.}

\item{...}{Further parameters passed to \code{[}.}

\item{j}{Column indices, names or logical vector.}
}
\description{
Cheaper alternative to \code{[} that consistently subsets data frame
rows, always returning a data frame. There are explicit methods for
enhanced data frames like tibbles, data.tables and sf.
}
\details{
\code{sset} is an S3 generic.
You can either write methods for \code{sset} or \code{[}. \cr
\code{sset} will fall back on using \code{[} when no suitable method is found.
}
\examples{
library(cheapr)
library(bench)
x <- rnorm(10^4)

mark(x[1:10^3], sset(x, 1:10^3))
mark(x[x > 0], sset(x, x > 0))

df <- data.frame(x = x)

mark(df[df$x > 0, , drop = FALSE],
     sset(df, df$x > 0),
     check = FALSE) # Row names are different

}
