% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste.R
\name{strings}
\alias{strings}
\alias{paste_}
\title{Fast string concatenation using C++}
\usage{
paste_(..., sep = "", collapse = NULL, .args = NULL)
}
\arguments{
\item{...}{Character vectors to concatenate.}

\item{sep}{\verb{[character(1)]} - A string to separate the supplied strings.}

\item{collapse}{Optional string to collapse concatenated strings into
one string (character vector of length 1).}

\item{.args}{An alternative to \code{...} so you can supply arguments directly
in a list. \cr
This is equivalent to \code{do.call(f, .args)} but much more efficient.}
}
\description{
Fast string concatenation using C++
}
\examples{
library(cheapr)
# Normal usage
paste_("Hello", "and", "Goodbye", sep = " ")
paste_(100, "\%")

paste_(letters, LETTERS)
paste_(letters, LETTERS, collapse = "")

# Both concatenating and collapsing
paste_(letters, LETTERS, sep = ",", collapse = " next letter ")
# This is the same as above
paste_(letters, LETTERS, collapse = ", next letter ")

# Recycling with zero-length vectors
paste_("hello", character(), letters)

# Concatenating many objects is very fast via `.args`
strings <- sample_(letters, 5e04, TRUE) |>
  with_local_seed(1) |>
  as.list()

strings <- lapply(strings, rep_len_, 2)
library(bench)
mark(
  paste_(.args = strings),
  do.call(paste0, strings)
)
mark(
  paste_(.args = strings, collapse = ""),
  do.call(paste0, c(strings, list(collapse = "")))
)

}
